+++ hijrical/__init__.py
@@ -78,7 +78,7 @@ from .tools import (
     upcoming_holidays,
 )
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
 
 __all__ = [
     # core